log results in __main__ by default
